fix: update async implementation
The original async implementation of this program hides all of the async
work behind a blocking call, which seems to mask the problem.  To force
this issue into the open, you need to await it on the main thread.  This
creates a Task that will complete when all of the supplied tasks have
completed and awaiting that array of Tasks exposes the problem.  If you
put it behing `Task.WaitAll`, the debugger seems to handle the blocking
call differently than the await.